Sort types working!
