+++ music_app/templates/results.html
@@ -1,9 +1,6 @@
 {% extends "search.html" %}
 {% block results %}
-<!-- This is all placeholder stuff. -->
 
-<!-- <h1>Enjoy this timeless music, if that's your thing or whatever.</h1>
-<br> -->
 {% if song_list %}
 <table>
   {% for song in song_list %}
